database interaction moved from Repository to databse [for proper structuring if there is a need to chnge database]
